fix: scoreConsecutives use size_t instead of unsigned

+++ src/common.h
@@ -88,7 +88,7 @@ extern Element ToUpper(const Element &s);
 extern bool isMatch(const CandidateString &subject, const Element &query_lw, const Element &query_up);
 extern bool isWordStart(const size_t pos, const CandidateString &subject, const CandidateString &subject_lw) noexcept;
 extern Score scoreCharacter(const size_t i, const size_t j, const bool start, const Score acro_score, const Score csc_score);
-extern Score scoreConsecutives(const CandidateString &subject, const CandidateString &subject_lw, const Element &query, const Element &query_lw, unsigned i, unsigned j, const bool startOfWord);
+extern Score scoreConsecutives(const CandidateString &subject, const CandidateString &subject_lw, const Element &query, const Element &query_lw, size_t i, size_t j, const bool startOfWord);
 extern AcronymResult scoreAcronyms(const CandidateString &subject, const CandidateString &subject_lw, const Element &query, const Element &query_lw);
 
 extern Score computeScore(const CandidateString &subject, const CandidateString &subject_lw, const PreparedQuery &preparedQuery);
